requirements.txt
altgraph==0.17.3
autopep8==2.0.2
numpy==1.24.2
pefile==2023.2.7
pycodestyle==2.10.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqtgraph==0.13.2
pyserial==3.5
pywin32-ctypes==0.2.0
